backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-ollama
ollama
pypdf
python-docx
